# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Plateforme d'Évaluation de la Maturité des DSIN
  version: 1.0.0
  description: Documentation API pour l'application d'évaluation de la maturité DevSecOps
  contact:
    name: Support API
    email: support@example.com
servers:
  - url: http://localhost:5000/api
    description: Serveur de développement
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ThematiqueScore:
      type: object
      properties:
        id_score:
          type: string
          format: uuid
          description: Identifiant unique du score
        thematique:
          type: string
          maxLength: 50
          description: Nom de la thématique
        score:
          type: number
          format: float
          description: Score pour cette thématique
        nombre_reponses:
          type: integer
          description: Nombre de réponses pour cette thématique
        id_analyse:
          type: string
          format: uuid
          description: Identifiant de l'analyse
      required:
        - id_analyse
        - thematique
        - score
        - nombre_reponses
      example:
        id_score: 550e8400-e29b-41d4-a716-446655440007
        id_analyse: 550e8400-e29b-41d4-a716-446655440006
        thematique: Sécurité
        score: 80.5
        nombre_reponses: 15
...
paths:
...
  /api/analyses/{idAnalyse}/thematiques:
    get:
      summary: Récupère les scores par thématique d'une analyse
      tags:
        - Thematiques
      parameters:
        - in: path
          name: idAnalyse
          required: true
          schema:
            type: string
            format: uuid
          description: Identifiant de l'analyse
      responses:
        '200':
          description: Scores par thématique récupérés avec succès
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ThematiqueScore'
        '404':
          description: Analyse non trouvée
        '500':
          description: Erreur serveur
...
  /api/stats/thematiques:
    get:
      summary: Récupère les statistiques par thématique
      tags:
        - Statistiques
      responses:
        '200':
          description: Statistiques par thématique récupérées avec succès
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    thematique:
                      type: string
                    scoreMoyen:
                      type: number
                      format: float
                    totalReponses:
                      type: integer
        '500':
          description: Erreur serveur
...
  /api/interpretation/thematiques/{idAnalyse}:
    get:
      summary: Récupère les interprétations par thématique pour une analyse
      tags:
        - Analyses
      parameters:
        - in: path
          name: idAnalyse
          required: true
          schema:
            type: string
            format: uuid
          description: Identifiant de l'analyse
      responses:
        '200':
          description: Interprétations par thématique récupérées avec succès
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    thematique:
                      type: string
                    score:
                      type: number
                      format: float
                    niveau:
                      type: string
                    description:
                      type: string
                    recommandations:
                      type: string
        '404':
          description: Analyse non trouvée
        '500':
          description: Erreur serveur
...